More accurate "plugged in" detection
We can end up plugged in, but no longer charging (f.g., if the device
has been plugged in for a while, it's obviously no longer charging, because
battery chemistry would go boom otherwise, but it's still plugged in).

For ref., on my H2O, online is 3 during charge, 1 in that "charged but
connected" state, and 0 otherwise.
TL;DR: For our purposes, we want it to be 0 ;).